+++ tosho/src/main.rs
@@ -34,6 +34,14 @@
 //!
 //! [![asciicast](https://asciinema.org/a/636303.svg)](https://asciinema.org/a/636303)
 //!
+//! ## Disclaimer
+//!
+//! This project is designed as an experiment and to create a local copy for personal use.
+//! These tools will not circumvent any paywall, and you will need to purchase and own each chapter on each platform
+//! with your own account to be able to make your own local copy.
+//!
+//! We're not responsible if your account got deactivated.
+//!
 //! ## License
 //!
 //! This project is licensed with MIT License ([LICENSE](https://github.com/noaione/tosho-mango/blob/master/LICENSE) or <http://opensource.org/licenses/MIT>)