+++ tosho_sjv/src/lib.rs
@@ -50,6 +50,14 @@
 //! }
 //! ```
 //!
+//! ## Disclaimer
+//!
+//! This project is designed as an experiment and to create a local copy for personal use.
+//! These tools will not circumvent any paywall, and you will need to purchase and own each chapter
+//! with your own account to be able to make your own local copy.
+//!
+//! We're not responsible if your account got deactivated.
+//!
 //! ## License
 //!
 //! This project is licensed with MIT License ([LICENSE](https://github.com/noaione/tosho-mango/blob/master/LICENSE) or <http://opensource.org/licenses/MIT>)