+++ tosho/src/main.rs
@@ -18,13 +18,13 @@
 //! Or...
 //!
 //! ```bash
-//! $ cargo install tosho
+//! cargo install --locked tosho
 //! ```
 //!
 //! Or, if you have [cargo-binstall](https://github.com/cargo-bins/cargo-binstall)...
 //!
 //! ```bash
-//! $ cargo binstall tosho
+//! cargo binstall --locked tosho
 //! ```
 //!
 //! ## Usage