+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Starting from Rust port of the project, all changes will be put into this file.
 
 ### Build
 - Change `diacritics` to `secular` because of OSS license issue.
-- Bump several CLI dependencies.
+- Bump dependencies on all crates (except `macros`)
 
 ## [0.4.2] (2024-02-28)
 ### Changes