+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 Starting from Rust port of the project, all changes will be put into this file.
 
 ## Unreleased (git master)
+
+Nothing yet!
+
+## [0.4.3] (2024-03-20)
 ### New Features
 - Check for update on startup (this is done for the next 24 hours after each check)